New Model Modifications
Some changes in the database structure were made. As such there is no longer a need for the Assessments model, and the Student model has had some alterations. The Studentclasses model has been changed and the subclasses model and it's logic has been added, so has the Standards model; which would the Schools to define their academic standards